devops-mvp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gitpython==3.1.40
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.4.2
aiofiles==23.2.1
pyyaml==6.0.1
